# docs/openapi.yaml
openapi: 3.0.3
info:
  title: tsz-go API
  description: |
    REST API for the tsz-go backend. All endpoints are JSON over HTTP.

    ## Authentication

    The API uses **JWT access tokens** + **refresh tokens**, stored differently
    by design:

    - **Access token** — returned in the JSON body (`access_token`). Lifetime 15
      minutes (default, `JWT_TTL`). Send it on protected endpoints via
      `Authorization: Bearer <access_token>`. Keep it in memory, not
      `localStorage`.
    - **Refresh token** — delivered as an HttpOnly, Secure, SameSite=Strict
      cookie (`refresh_token`, scoped to `Path=/api/v1/auth`), **not** in the
      body. Lifetime 30 days (default, `REFRESH_TOKEN_TTL`). The browser sends it
      automatically to `/auth/refresh` and `/auth/logout`. Make requests with
      credentials (`withCredentials: true` / `credentials: 'include'`).

    Login is **strict single-device**: each successful login/register revokes the
    user's previous sessions, and using the refresh cookie rotates it.

    See `docs/api.md` for the full rationale, 401-handling decision tree, and
    frontend integration guide.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/me/learning-settings:
    put:
      tags: [User]
      summary: Set the learner's CEFR level + English variant
      operationId: updateLearningSettings
      description: |
        Sets the two basic learning settings. Backs both new-user onboarding and
        later edits from the settings screen. Both fields are required and written
        together, so "onboarded" is never half-set.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LearningSettingsRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningSettingsResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthError"
        "409":
          description: Conflict — the user has no student profile to attach settings to.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { error: "learning settings require a student profile" }
  /api/v1/auth/logout-all:
    post:
      tags: [Auth (authenticated)]
      summary: Revoke every refresh token (logout everywhere)
      operationId: logoutAll
      description: |
        Revokes every refresh token the authenticated user holds, driven by the
        access token's subject (no refresh cookie needed). Idempotent.
      responses:
        "204":
          description: No Content (also if the user has no active sessions).
        "401":
          $ref: "#/components/responses/AuthError"
  /api/v1/auth/switch-role:
    post:
      tags: [Auth (authenticated)]
      summary: Switch active role to one the user already holds
      operationId: switchRole
      description: Re-issues an access token scoped to the role. Refresh token unchanged.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleTokenResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthError"
        "403":
          description: Forbidden — user does not have this role.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { error: "user does not have this role" }
  /api/v1/auth/roles:
    post:
      tags: [Auth (authenticated)]
      summary: Acquire an additional role, then switch to it
      operationId: addRole
      description: |
        Grants the user a second identity (e.g. a student who also teaches) and
        returns an access token scoped to the new role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleTokenResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthError"
        "409":
          description: Conflict — user already has this role.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { error: "user already has this role" }

  # ── Admin (back office) ──────────────────────────────────────────────
  # All /api/v1/admin/* endpoints require an access token whose active role is
  # admin; the admin middleware returns 403 otherwise. Design-only here — no
  # handlers are generated for these (apigen is scoped to getCurrentUser).
...
  /api/v1/admin/profile:
    get:
      tags: [Admin (users)]
      summary: Current admin's own identity
      operationId: adminProfile
      description: |
        Phase A probe behind the admin gate. Lets the console confirm the role
        gate (200 for admin, 403 otherwise) and render a "signed in as …" header.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfile"
        "401":
          $ref: "#/components/responses/AuthError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token: `Authorization: Bearer <access_token>`."
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HttpOnly refresh-token cookie, sent automatically by the browser.
...
  responses:
    ValidationError:
      description: Bad Request — the body failed validation. `error` carries the raw validator message.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    AuthError:
      description: |
        Unauthorized. `missing or malformed authorization header` (header absent
        or not `Bearer <token>`) or `invalid or expired token` (access token
        expired/tampered — call `/auth/refresh`, then retry).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            missing:
              value: { error: "missing or malformed authorization header" }
            expired:
              value: { error: "invalid or expired token" }
    ForbiddenError:
      description: Forbidden — the access token's active role is not admin.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { error: "admin role required" }
    NotFoundError:
      description: Not Found — the referenced resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { error: "not found" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable message.
      required: [error]
    Role:
      type: string
      enum: [student, teacher]
    CEFRLevel:
      type: string
      enum: [A1, A2, B1, B2, C1, C2]
      description: Learner difficulty band (CEFR), A1 easiest … C2 hardest.
    EnglishVariant:
      type: string
      enum: [BrE, AmE]
      description: Accent and spelling convention — British or American.
    LearningSettings:
      type: object
      description: A learner's two onboarding choices; both are always set together.
      properties:
        cefr_level:
          $ref: "#/components/schemas/CEFRLevel"
        english_variant:
          $ref: "#/components/schemas/EnglishVariant"
      required: [cefr_level, english_variant]
    LearningSettingsRequest:
      type: object
      description: Both fields required — the two basic settings are written together.
      required: [cefr_level, english_variant]
      properties:
        cefr_level:
          $ref: "#/components/schemas/CEFRLevel"
        english_variant:
          $ref: "#/components/schemas/EnglishVariant"
...
    RoleRequest:
      type: object
      required: [role]
      properties:
        role:
          $ref: "#/components/schemas/Role"
...
    LearningSettingsResponse:
      type: object
      description: Returned by update-learning-settings.
      properties:
        learning_settings:
          $ref: "#/components/schemas/LearningSettings"
        onboarded:
          type: boolean
          description: Always true here — settings were just written.
          example: true
      required: [learning_settings, onboarded]
    RoleTokenResponse:
      type: object
      description: Returned by switch-role / add-role.
      properties:
        access_token:
          type: string
          example: "jwt..."
        active_role:
          $ref: "#/components/schemas/Role"
      required: [access_token, active_role]
...
    AdminProfile:
      type: object
      description: The signed-in admin's own identity (Phase A probe).
      properties:
        id:
          type: string
          format: uuid
        phone:
          type: string
          example: "13800138000"
        display_name:
          type: string
          example: Administrator
        roles:
          type: array
          items:
            type: string
          description: All roles the account holds (includes admin).
          example: [admin]
        active_role:
          type: string
          description: The role this access token is acting as.
          example: admin
      required: [id, phone, display_name, roles, active_role]
...
security:
  - bearerAuth: []